requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-core>=0.2.38
langchain-community>=0.2.16
langchain-openai>=0.0.5
langchain-chroma>=0.1.3
langchain-text-splitters>=0.2.4

# Vector store and embeddings
chromadb>=0.4.0
chroma-hnswlib>=0.7.3

# Document processing
unstructured>=0.12.0
pdf2image>=1.16.3
pytesseract>=0.3.10
# poppler-utils>=0.68.0
poppler-utils
pdfminer.six>=20221105
pdfplumber>=0.10.3
pypdf>=4.3.1
Pillow>=10.0.0

# LLM and API
openai>=1.1.0
tiktoken>=0.5.0

# Data handling and utilities
pandas>=2.0.0
numpy>=1.26.4
PyYAML>=6.0.2
python-dotenv>=1.0.0
pyprojroot>=0.3.0

# Database
sqlalchemy>=2.0.0

# Type checking and validation
pydantic>=2.9.0
pydantic-core>=2.23.2

# Audio processing
pydub>=0.25.1
pyaudioop>=0.2.0
soundfile>=0.12.1

# Web interface
gradio>=4.43.0
ffmpeg-python>=0.2.0

# Optional - for Jupyter notebooks
ipykernel>=6.29.5
jupyter>=1.0.0
notebook>=7.0.0

# Web interface (if needed)
gradio>=4.43.0

audioop-lts
portaudio
pyaudio